Split mapping definitions
